+++ tests/11-orm.ts
@@ -1,4 +1,4 @@
-import { nSQL } from "../src/index";
+import { NanoSQLInstance } from "../src/index";
 import { expect, assert } from "chai";
 import "mocha";
 import { usersDB, ExampleUsers, ExampleDataModel } from "./data";
@@ -17,24 +17,25 @@ const ormDataModels = {
 
 describe("ORM System", () => {
     it("Related records should auto add.", (done: MochaDone) => {
-        nSQL("users").model(ormDataModels.users.concat([
+        const nSQL = new NanoSQLInstance();
+        nSQL.table("users").model(ormDataModels.users.concat([
             {key: "posts", type: "posts[]", props: ["ref=>author"] }
         ]));
-        nSQL("posts").model(ormDataModels.posts.concat([
+        nSQL.table("posts").model(ormDataModels.posts.concat([
             {key: "author", type: "users", props: ["ref=>posts[]"]}
         ]));
-        nSQL().connect().then(() => {
-            return nSQL("users").query("upsert", {
+        nSQL.connect().then(() => {
+            return nSQL.table("users").query("upsert", {
                 name: "billy",
                 posts: []
             }).exec();
         }).then(() => {
-            return nSQL("posts").query("upsert", {
+            return nSQL.table("posts").query("upsert", {
                 title: "How To Fly",
                 author: 1
             }).exec();
         }).then(() => {
-            return nSQL("users").query("select").exec();
+            return nSQL.table("users").query("select").exec();
         }).then((rows) => {
             try {
                 expect(rows).to.deep.equal( [ { name: "billy", posts: [ 1 ], id: 1 } ], "Auto add ORM failed!");
@@ -45,29 +46,30 @@ describe("ORM System", () => {
         });
     });
     it("Related records should auto remove when updated.", (done: MochaDone) => {
-        nSQL("users").model(ormDataModels.users.concat([
+        const nSQL = new NanoSQLInstance();
+        nSQL.table("users").model(ormDataModels.users.concat([
             {key: "posts", type: "posts[]", props: ["ref=>author"] }
         ]));
-        nSQL("posts").model(ormDataModels.posts.concat([
+        nSQL.table("posts").model(ormDataModels.posts.concat([
             {key: "author", type: "users", props: ["ref=>posts[]"]}
         ]));
-        nSQL().connect().then(() => {
-            return nSQL("users").query("upsert", {
+        nSQL.connect().then(() => {
+            return nSQL.table("users").query("upsert", {
                 name: "billy",
                 posts: []
             }).exec();
         }).then(() => {
-            return nSQL("posts").query("upsert", {
+            return nSQL.table("posts").query("upsert", {
                 title: "How To Fly",
                 author: 1
             }).exec();
         }).then(() => {
-            return nSQL("posts").query("upsert", {
+            return nSQL.table("posts").query("upsert", {
                 id: 1,
                 author: null
             }).exec();
         }).then(() => {
-            return nSQL("users").query("select").exec();
+            return nSQL.table("users").query("select").exec();
         }).then((rows) => {
             try {
                 expect(rows).to.deep.equal( [ { name: "billy", posts: [], id: 1 } ], "Auto remove on update failed!");
@@ -78,29 +80,30 @@ describe("ORM System", () => {
         });
     });
     it("Related records should auto remove when updated (Other direction).", (done: MochaDone) => {
-        nSQL("users").model(ormDataModels.users.concat([
+        const nSQL = new NanoSQLInstance();
+        nSQL.table("users").model(ormDataModels.users.concat([
             {key: "posts", type: "posts[]", props: ["ref=>author"] }
         ]));
-        nSQL("posts").model(ormDataModels.posts.concat([
+        nSQL.table("posts").model(ormDataModels.posts.concat([
             {key: "author", type: "users", props: ["ref=>posts[]"]}
         ]));
-        nSQL().connect().then(() => {
-            return nSQL("users").query("upsert", {
+        nSQL.connect().then(() => {
+            return nSQL.table("users").query("upsert", {
                 name: "billy",
                 posts: []
             }).exec();
         }).then(() => {
-            return nSQL("posts").query("upsert", {
+            return nSQL.table("posts").query("upsert", {
                 title: "How To Fly",
                 author: 1
             }).exec();
         }).then(() => {
-            return nSQL("users").query("upsert", {
+            return nSQL.table("users").query("upsert", {
                 id: 1,
                 posts: []
             }).exec();
         }).then(() => {
-            return nSQL("posts").query("select").exec();
+            return nSQL.table("posts").query("select").exec();
         }).then((rows) => {
             try {
                 expect(rows).to.deep.equal( [ { title: "How To Fly", author: null, id: 1 } ], "Auto remove on update failed!");
@@ -112,26 +115,27 @@ describe("ORM System", () => {
     });
 
     it("Related records should auto remove when deleted.", (done: MochaDone) => {
-        nSQL("users").model(ormDataModels.users.concat([
+        const nSQL = new NanoSQLInstance();
+        nSQL.table("users").model(ormDataModels.users.concat([
             {key: "posts", type: "posts[]", props: ["ref=>author"] }
         ]));
-        nSQL("posts").model(ormDataModels.posts.concat([
+        nSQL.table("posts").model(ormDataModels.posts.concat([
             {key: "author", type: "users", props: ["ref=>posts[]"]}
         ]));
-        nSQL().connect().then(() => {
-            return nSQL("users").query("upsert", {
+        nSQL.connect().then(() => {
+            return nSQL.table("users").query("upsert", {
                 name: "billy",
                 posts: []
             }).exec();
         }).then(() => {
-            return nSQL("posts").query("upsert", {
+            return nSQL.table("posts").query("upsert", {
                 title: "How To Fly",
                 author: 1
             }).exec();
         }).then(() => {
-            return nSQL("posts").query("delete").where(["id", "=", 1]).exec();
+            return nSQL.table("posts").query("delete").where(["id", "=", 1]).exec();
         }).then(() => {
-            return nSQL("users").query("select").exec();
+            return nSQL.table("users").query("select").exec();
         }).then((rows) => {
             try {
                 expect(rows).to.deep.equal( [ { name: "billy", posts: [], id: 1 } ], "Auto remove on delete failed!");
@@ -143,24 +147,25 @@ describe("ORM System", () => {
     });
 
     it("Many to Many relationships should remain in sync.", (done: MochaDone) => {
-        nSQL("users").model(ormDataModels.users.concat([
+        const nSQL = new NanoSQLInstance();
+        nSQL.table("users").model(ormDataModels.users.concat([
             {key: "posts", type: "posts[]", props: ["ref=>authors[]"] }
         ]));
-        nSQL("posts").model(ormDataModels.posts.concat([
+        nSQL.table("posts").model(ormDataModels.posts.concat([
             {key: "authors", type: "users[]", props: ["ref=>posts[]"]}
         ]));
-        nSQL().connect().then(() => {
-            return nSQL("users").loadJS("users", [
+        nSQL.connect().then(() => {
+            return nSQL.table("users").loadJS("users", [
                 {name: "Billy", posts: []},
                 {name: "Jeb", posts: []}
             ]);
         }).then(() => {
-            return nSQL("posts").query("upsert", {
+            return nSQL.table("posts").query("upsert", {
                 title: "How To Fly",
                 authors: [1, 2]
             }).exec();
         }).then(() => {
-            return nSQL("users").query("select").exec();
+            return nSQL.table("users").query("select").exec();
         }).then((rows) => {
             try {
                 expect(rows).to.deep.equal( [
@@ -175,26 +180,27 @@ describe("ORM System", () => {
     });
 
     it("Many to Many relationships should remain in sync when a delete happens.", (done: MochaDone) => {
-        nSQL("users").model(ormDataModels.users.concat([
+        const nSQL = new NanoSQLInstance();
+        nSQL.table("users").model(ormDataModels.users.concat([
             {key: "posts", type: "posts[]", props: ["ref=>authors[]"] }
         ]));
-        nSQL("posts").model(ormDataModels.posts.concat([
+        nSQL.table("posts").model(ormDataModels.posts.concat([
             {key: "authors", type: "users[]", props: ["ref=>posts[]"]}
         ]));
-        nSQL().connect().then(() => {
-            return nSQL("users").loadJS("users", [
+        nSQL.connect().then(() => {
+            return nSQL.table("users").loadJS("users", [
                 {name: "Billy", posts: []},
                 {name: "Jeb", posts: []}
             ]);
         }).then(() => {
-            return nSQL("posts").query("upsert", {
+            return nSQL.table("posts").query("upsert", {
                 title: "How To Fly",
                 authors: [1, 2]
             }).exec();
         }).then(() => {
-            return nSQL("posts").query("delete").where(["id", "=", 1]).exec();
+            return nSQL.table("posts").query("delete").where(["id", "=", 1]).exec();
         }).then(() => {
-            return nSQL("users").query("select").exec();
+            return nSQL.table("users").query("select").exec();
         }).then((rows) => {
             try {
                 expect(rows).to.deep.equal( [
@@ -209,26 +215,27 @@ describe("ORM System", () => {
     });
 
     it("Many to Many relationships should remain in sync when a delete happens (other direction).", (done: MochaDone) => {
-        nSQL("users").model(ormDataModels.users.concat([
+        const nSQL = new NanoSQLInstance();
+        nSQL.table("users").model(ormDataModels.users.concat([
             {key: "posts", type: "posts[]", props: ["ref=>authors[]"] }
         ]));
-        nSQL("posts").model(ormDataModels.posts.concat([
+        nSQL.table("posts").model(ormDataModels.posts.concat([
             {key: "authors", type: "users[]", props: ["ref=>posts[]"]}
         ]));
-        nSQL().connect().then(() => {
-            return nSQL("users").loadJS("users", [
+        nSQL.connect().then(() => {
+            return nSQL.table("users").loadJS("users", [
                 {name: "Billy", posts: []},
                 {name: "Jeb", posts: []}
             ]);
         }).then(() => {
-            return nSQL("posts").query("upsert", {
+            return nSQL.table("posts").query("upsert", {
                 title: "How To Fly",
                 authors: [1, 2]
             }).exec();
         }).then(() => {
-            return nSQL("users").query("delete").where(["id", "=", 1]).exec();
+            return nSQL.table("users").query("delete").where(["id", "=", 1]).exec();
         }).then(() => {
-            return nSQL("posts").query("select").exec();
+            return nSQL.table("posts").query("select").exec();
         }).then((rows) => {
             try {
                 expect(rows).to.deep.equal( [
@@ -242,23 +249,24 @@ describe("ORM System", () => {
     });
 
     it("ORM Query.", (done: MochaDone) => {
-        nSQL("users").model(ormDataModels.users);
-        nSQL("posts").model(ormDataModels.posts.concat([
+        const nSQL = new NanoSQLInstance();
+        nSQL.table("users").model(ormDataModels.users);
+        nSQL.table("posts").model(ormDataModels.posts.concat([
             {key: "authors", type: "users[]"}
         ]));
-        nSQL().connect().then(() => {
-            return nSQL("users").loadJS("users", [
+        nSQL.connect().then(() => {
+            return nSQL.table("users").loadJS("users", [
                 {name: "Billy"},
                 {name: "Jeb"},
                 {name: "Val"}
             ]);
         }).then(() => {
-            return nSQL("posts").query("upsert", {
+            return nSQL.table("posts").query("upsert", {
                 title: "How To Fly",
                 authors: [1, 3]
             }).exec();
         }).then(() => {
-            return nSQL("posts").query("select").orm(["authors"]).exec();
+            return nSQL.table("posts").query("select").orm(["authors"]).exec();
         }).then((rows) => {
             try {
                 expect(rows).to.deep.equal( [