+++ tests/12-events.ts
@@ -1,4 +1,3 @@
-import { nSQL } from "../src/index";
 import { expect, assert } from "chai";
 import "mocha";
 import { usersDB, ExampleUsers, ExampleDataModel } from "./data";
@@ -6,6 +5,7 @@ import { usersDB, ExampleUsers, ExampleDataModel } from "./data";
 describe("Events", () => {
     it("Event Listener should trigger.", (done: MochaDone) => {
         usersDB(ExampleDataModel, (nSQL) => {
+
             nSQL.loadJS("users", ExampleUsers).then(() => {
 
                 const onSelect = (event) => {